requirements.txt
requests==2.32.5
structlog==25.5.0
tenacity==9.1.2
pyyaml==6.0.3
pandas==2.3.3
numpy==2.2.6
